pyproject.toml
[tool.poetry]
name = "onboarding-assignment"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flake8 = "^3.9.2"
black = "^21.9b0"
mypy = "^0.910"
httpx = "^0.19.0"
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
